Working update 2.5
Optimizations
