+++ .gitignore
@@ -1,0 +1,6 @@
+# Folders
+cgi-bin/__pycache__/
+BACKUP/
+.vscode/
+
+
