+++ .htaccess
@@ -1,0 +1,10 @@
+RewriteEngine on
+RewriteBase /
+
+RewriteCond %{REQUEST_URI} !^/cgi-bin/
+
+# Rewrites all URLS [Replace "example" with the actual domain, without the TLD (.com, .net, .biz, etc)]
+RewriteCond %{HTTP_HOST} ^(www\.)?wadaily\.
+
+# Rewrite all those to insert /folder
+RewriteRule ^(.*)$ /cgi-bin/code.py/$1 [L]
