+++ cgi-bin/ScheduleCreator.html
@@ -1,0 +1,31 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <!--
+                          Devloped by Trey W. 
+  __      __  ______  ____                   ___
+ /\ \  __/\ \/\  _  \/\  _`\             __ /\_ \
+ \ \ \/\ \ \ \ \ \L\ \ \ \/\ \     __   /\_\\//\ \    __  __
+  \ \ \ \ \ \ \ \  __ \ \ \ \ \  /'__`\ \/\ \ \ \ \  /\ \/\ \
+   \ \ \_/ \_\ \ \ \/\ \ \ \_\ \/\ \L\.\_\ \ \ \_\ \_\ \ \_\ \
+    \ `\___x___/\ \_\ \_\ \____/\ \__/.\_\\ \_\/\____\\/`____ \
+     '\/__//__/  \/_/\/_/\/___/  \/__/\/_/ \/_/\/____/ `/___/> \
+                                                          /\___/
+                                                          \/__/
+                          Your daily updates
+  -->
+  <title>Schedule Generator</title>
+</head>
+<body>
+  <form action="/action_page.php">
+  First name:<br>
+  <input type="text" name="firstname" value="Mickey">
+  <br>
+  Last name:<br>
+  <input type="text" name="lastname" value="Mouse">
+  <br><br>
+  <input type="submit" value="Submit">
+</form>
+  <h1>Test</h1>
+</body>
+</html>
