+++ cgi-bin/ScheduleDictStorage.txt
@@ -1,0 +1,1 @@
+{"May 16 2019": "DAY 9", "May 13 2019": "DAY 6", "May 14 2019": "DAY 7", "May 15 2019": "DAY 8", "May 9 2019": "DAY 4", "May 10 2019": "DAY 5", "May 6 2019": "DAY 1", "May 7 2019": "DAY 2", "May 8 2019": "DAY 3", "May 2 2019": "SPRING", "May 3 2019": "DAY 7", "April 29 2019": "DAY 4", "April 30 2019": "DAY 5", "May 1 2019": "DAY 6", "April 25 2019": "DAY 2", "April 26 2019": "DAY 3", "April 23 2019": "XDAY", "April 24 2019": "DAY 1", "April 18 2019": "XDAY", "April 15 2019": "DAY 7", "April 16 2019": "DAY 8", "April 17 2018": "DAY 9", "April 11 2019": "DAY 5", "April 12 2019": "DAY 6", "April 8 2019": "DAY 3", "April 9 2019": "DAY 4", "April 10 2019": "XDAY", "April 4 2019": "DAY 1", "April 5 2019": "DAY 2", "April 2 2019": "DAY 9", "April 1 2019": "DAY 8", "April 3 2019": "XDAY", "March 28 2019": "DAY 6", "March 29 2019": "DAY 7", "March 25 2019": "DAY 3", "March 26 2019": "DAY 4", "March 27 2019": "DAY 5", "March 22 2019": "DAY 2", "March 21 2019": "DAY 1", "March 18 2019": "DAY 5", "March 19 2019": "DAY 6", "March 20 2019": "DAY 7", "March 14 2019": "DAY 3", "March 15 2019": "DAY 4", "March 12 2019": "DAY 1", "March 13 2019": "DAY 2", "February 28 2019": "DAY 6", "March 1 2019": "DAY 7", "February 25 2019": "DAY 3", "February 26 2019": "DAY 4", "February 27 2019": "DAY 5", "February 21 2019": "DAY 1", "February 22 2019": "DAY 2", "February 19 2019": "DAY 8", "February 20 2019": "DAY 9", "February 14 2019": "DAY 7", "February 15 2019": "XDAY", "February 11 2019": "DAY 4", "February 12 2019": "DAY 5", "February 13 2019": "DAY 6", "February 7 2019": "DAY 2", "February 8 2019": "DAY 3", "February 4 2019": "DAY 8", "February 5 2019": "DAY 9", "February 6 2019": "DAY 1", "January 31 2019": "DAY 6", "February 1 2019": "DAY 7", "January 28 2019": "DAY 3", "January 29 2019": "DAY 4", "January 30 2019": "DAY 5", "January 24 2019": "DAY 1", "January 25 2019": "DAY 2", "22 2019 XDAY": "XDAY", "23 2019 A": "XDAY", "January 17 2019": "DAY 8", "January 18 2019": "DAY 9", "January 14 2019": "DAY 5", "January 15 2019": "DAY 6", "January 16 2019": "DAY 7", "January 10 2019": "DAY 3", "January 11 2019": "DAY 4", "January 8 2019": "DAY 1", "January 9 2019": "DAY 2", "December 13 2018": "DAY 9", "December 10 2018": "DAY 8", "December 11 2018": "DAY 9", "December 12 2018": "DAY 8", "December 6 2018": "DAY 7", "December 7 2018": "XDAY", "December 3 2018": "DAY 4", "December 4 2018": "DAY 5", "December 5 2018": "DAY 6", "November 29 2018": "DAY 2", "November 30 2018": "DAY 3", "November 26 2018": "DAY 8", "November 27 2018": "DAY 9", "November 28 2018": "DAY 1", "November 15 2018": "DAY 8", "November 16 2018": "DAY 9", "November 12 2018": "DAY 5", "November 13 2018": "DAY 6", "November 14 2018": "DAY 7", "November 8 2018": "DAY 3", "November 9 2018": "DAY 4", "November 5 2018": "DAY 1", "November 6 2018": "DAY 2", "November 1 2018": "DAY 8", "November 2 2018": "DAY 9", "October 29 2018": "DAY 5", "October 30 2018": "DAY 6", "October 31 2018": "DAY 7", "October 25 2018": "DAY 3", "October 26 2018": "DAY 4", "October 22 2018": "DAY 9", "October 23 2018": "DAY 1", "October 24 2018": "DAY 2", "October 18 2018": "DAY 8", "October 19 2018": "XDAY", "October 15 2018": "DAY 5", "October 16 2018": "DAY 6", "October 17 2018": "DAY 7", "October 11 2018": "DAY 3", "October 12 2018": "DAY 4", "October 8 2018": "DAY 1", "October 9 2018": "DAY 2", "October 10 2018": "SPECIAL", "October 4 2018": "DAY 8", "October 5 2018": "DAY 9", "October 1 2018": "DAY 5", "October 2 2018": "DAY 6", "October 3 2018": "DAY 7", "September 27 2018": "DAY 3", "September 28 2018": "DAY 4", "September 24 2018": "DAY 9", "September 25 2018": "DAY 1", "September 26 2018": "DAY 2", "September 21 2018": "DAY 8", "September 20 2018": "XDAY", "September 17 2018": "DAY 6", "September 18 2018": "DAY 7", "September 13 2018": "MID DAY ASM", "September 14 2018": "DAY 5", "September 11 2018": "DAY 2", "September 12 2018": "DAY 3", "September 6 2018": "DAY 9", "September 7 2018": "DAY 1", "September 4 2018": "DAY 7", "September 5 2018": "DAY 8", "August 30 2018": "DAY 5", "August 31 2018": "DAY 6", "August 27 2018": "DAY 2", "August 28 2018": "DAY 3", "August 12 2018": "DAY 1", "August 29 2018": "DAY 4", "August 23 2018": "DAY 9", "August 24 2018": "DAY 1", "August 20 2018": "DAY 6", "August 21 2018": "DAY 7", "August 22 2018": "DAY 8", "August 16 2018": "DAY 4", "August 17 2018": "DAY 5", "August 13 2018": "DAY 1", "August 14 2018": "DAY 2", "August 15 2018": "DAY 3"}
