+++ cgi-bin/lunchList.txt
@@ -1,0 +1,1 @@
+{"Tuesday": ["Jamaican Jerk Pork w/ Pineapple Salsa", "Oven Baked BBQ Chicken", "Jamaican Vegetable Patties", "Caribbean Rice", "Steamed Napa Cabbage", "Hawaiian Dinner Roll", "Corn & Summer Vegetable Chowder", "West Commons: Cuban Sandwich, Jamaican Vegetable Patties, Caesar or Italian Salad"]}{"Friday": ["Jamaican Jerk Pork w/ Pineapple Salsa", "Oven Baked BBQ Chicken", "Jamaican Vegetable Patties", "Caribbean Rice", "Steamed Napa Cabbage", "Hawaiian Dinner Roll", "Corn & Summer Vegetable Chowder", "West Commons: Cuban Sandwich, Jamaican Vegetable Patties, Caesar or Italian Salad"]}
