+++ cgi-bin/templates/index.html
@@ -1,0 +1,135 @@
+$def with (rotation, weekday, foodList, image, events, date)
+<!-- Developed by Trey W. -->
+<!DOCTYPE html>
+<html>
+<head>
+  <title>WADaily</title>
+  <link rel="shortcut icon" href="../images/favicon.ico" type="image/x-icon"></link>
+  <link href="https://fonts.googleapis.com/icon?family=Material+Icons"
+      rel="stylesheet"></link>
+  <style>
+    body {
+      text-align: center;
+      font-weight: 300;
+      color: #696a6d;
+      font-style: normal;
+    }
+    h2 {
+      font-family: "crimson", sans-serif;
+      font-weight: 500;
+      font-size: 41px;
+      line-height: 0.65;
+      margin-bottom: 21px;
+      margin-top: 21px;
+      color: #a32136;
+    }
+    #scheduleLeft {margin-right:10px;position:relative;bottom:$image['bottomPadding'];}
+    #scheduleImage {display:inline-block;position:relative;}
+    #scheduleRight {margin-left:10px;position:relative;bottom:$image['bottomPadding'];}
+    #refresh {margin-top:10px}
+    .material-icons.md-dark {color: rgba(0, 0, 0, 1);}
+    .material-icons.md-36 {font-size:38px;width:38px;height:38px;}
+    .material-icons.md-darkish {color: rgba(0, 0, 0, 0.75)}
+    .button .md-dark, .button:hover .md-darkish {display: none;}
+    .button:hover .md-dark {display: inline-block;text-shadow:1px 1px 1.5px rgba(0,0,0,0.5);}
+    .disabled {opacity:0.6;cursor:not-allowed;}
+    ul   {text-align: center;}
+  </style>
+  <script async src="https://www.googletagmanager.com/gtag/js?id=UA-119002335-1"></script>
+</head>
+
+<body>
+  <div class="Rotation">
+    <h2>Rotation</h2>
+        $if rotation:
+            It is <b>$weekday</b> and it's <b>$rotation</b> rotation!<br/>
+            <p id="date">$date[0] $date[1] $date[2]</p>
+        $else:
+            Could not find the <b>Rotation</b>!<br/>
+            <p id="date">$date[0] $date[1] $date[2]</p>
+      <div>
+      <a = href="#" onclick="plusDay(-1)"class="button">
+        <i id="scheduleLeft" class="material-icons md-36 md-darkish">arrow_backward</i>
+        <i id="scheduleLeft" class="material-icons md-36 md-dark">arrow_backward</i>
+      </a>
+      $if image['src']:
+          <img id="scheduleImage" src=$image['src'] alt=$image['alt'] width=$image['width'] height=$image['height']>
+      $else:
+          <b style="display:inline-block;margin-left:30px;margin-right:30px;">$weekday</b>
+      <a href="#" class="button" onclick="plusDay(1)">
+        <i id="scheduleRight" class="material-icons md-36 md-darkish">arrow_forward</i>
+        <i id="scheduleRight" class="material-icons md-36 md-dark">arrow_forward</i>
+      </a>
+      </div>
+      <a href="http://127.0.0.1:8080/" class="button">
+        <i id="refresh" class="material-icons md-36 md-darkish">refresh</i>
+        <i id="refresh" class="material-icons md-36 md-dark">refresh</i>
+      </a>
+  </div>
+  <div class="Lunch">
+    <h2>Lunch Menu</h2>
+    $if foodList:
+        <p>
+        $for food in foodList:
+            $food<br/>
+        </p>
+    $else:
+        Could not find the <b>lunch</b>!
+  </div>
+  <div class="Events">
+    <h2>Today's Events</h2>
+    $if events:
+        <p>
+        $for y in events:
+            $y<br/>
+        </p>
+    $else:
+        <p>No events.</p>
+  </div>
+    <div class="Footer">
+    <h6>Developed by <b>Trey W.</b><br>
+    This website may have some affiliation with Woodward Academy</h6>
+  </div>
+</body>
+
+<script>
+  window.dataLayer = window.dataLayer || [];
+  function gtag(){dataLayer.push(arguments);}
+  gtag('js', new Date());
+
+  gtag('config', 'UA-119002335-1');
+
+  Date.prototype.addDays = function(days, foo) {
+      if (foo == true){
+        var date = new Date(this.valueOf());
+      }else{
+        var date = new Date(foo);
+      }
+      date.setDate(date.getDate() + days);
+      return date;
+  }
+
+  var dayIndex = 0;
+  function formatDate(date) {
+    var monthNames = [
+      "January", "February", "March",
+      "April", "May", "June", "July",
+      "August", "September", "October",
+      "November", "December"
+    ];
+
+    var day = date.getDate();
+    var monthIndex = date.getMonth();
+    var year = date.getFullYear();
+
+    return [(monthNames[monthIndex] + ' ' + day + ' ' + year), ('http://127.0.0.1:8080/?date='+ monthNames[monthIndex] + '%20' + day + '%20' + year)];
+  }
+
+  var date = new Date();
+  var updatedDate = document.getElementById('date').innerHTML;
+
+  function plusDay(n) {
+    window.location = formatDate(date.addDays((dayIndex = n), (updatedDate)))[1]
+  }
+</script>
+</html>
